requirements.txt
anyio==3.6.1
autopep8==1.7.0
click==8.1.3
fastapi==0.79.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pydantic==1.9.1
python-decouple==3.6
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing-extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
